refactor(frontend): move AlbumPageLayout to organisms to fix atomic-design layer
AlbumPageLayout imported PlaylistTracksList (an organism), so a molecule
depended on an organism — an upward atomic-design violation. The component is
organism-level by composition; relocate it and fix the relative imports.

- Move components/molecules/AlbumPageLayout.{tsx,test.tsx} -> components/organisms/.
- PlaylistTracksList import becomes sibling (./); EmptyState/PlaylistHeader
  become ../molecules/.
- Update consumers: organisms/Playlist.tsx (+test), views/LibraryAlbumDetail.tsx.

Pure relocation, no behavior change. Frontend lint/build pass, 495 tests green.
