+++ apps/frontend/src/components/organisms/AlbumPageLayout.test.tsx
